Minor tweaking, renaming
