+++ src/test/java/com/fasterxml/jackson/core/constraints/AsyncLargeNumberReadTest.java
@@ -1,0 +1,107 @@
+package com.fasterxml.jackson.core.constraints;
+
+import com.fasterxml.jackson.core.async.ByteArrayFeeder;
+import org.junit.jupiter.api.Test;
+
+import com.fasterxml.jackson.core.*;
+import com.fasterxml.jackson.core.exc.StreamConstraintsException;
+import com.fasterxml.jackson.core.JsonFactory;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * Number Length Constraint Bypass in Non-Blocking (Async) JSON Parsers
+ */
+class AsyncLargeNumberReadTest
+    extends com.fasterxml.jackson.core.JUnit5TestBase
+{
+    private static final int TEST_NUMBER_LENGTH = StreamReadConstraints.DEFAULT_MAX_NUM_LEN * 2;
+
+    private final JsonFactory JSON_F = newStreamFactory();
+
+    @Test
+    void asyncParserFailsTooLongInt() throws Exception {
+        byte[] payload = buildPayloadWithLongInteger(TEST_NUMBER_LENGTH);
+
+        try (JsonParser p = JSON_F.createNonBlockingByteArrayParser()) {
+            ByteArrayFeeder byteArrayFeeder = (ByteArrayFeeder) p;
+            byteArrayFeeder.feedInput(payload, 0, payload.length);
+            byteArrayFeeder.endOfInput();
+    
+            _asyncParserFailsTooLongNumber(p, JsonToken.VALUE_NUMBER_INT);
+        }
+    }
+
+    @Test
+    void asyncParserFailsTooLongDecimal() throws Exception {
+        byte[] payload = buildPayloadWithLongDecimal(TEST_NUMBER_LENGTH);
+
+        try (JsonParser p = JSON_F.createNonBlockingByteArrayParser()) {
+            ByteArrayFeeder byteArrayFeeder = (ByteArrayFeeder) p;
+            byteArrayFeeder.feedInput(payload, 0, payload.length);
+            byteArrayFeeder.endOfInput();
+
+            _asyncParserFailsTooLongNumber(p, JsonToken.VALUE_NUMBER_FLOAT);
+        }
+    }
+
+    @Test
+    void asyncParserFailsTooLongDecimalWithExponent() throws Exception {
+        byte[] payload = buildPayloadWithLongExponent(TEST_NUMBER_LENGTH);
+
+        try (JsonParser p = JSON_F.createNonBlockingByteArrayParser()) {
+            ByteArrayFeeder byteArrayFeeder = (ByteArrayFeeder) p;
+            byteArrayFeeder.feedInput(payload, 0, payload.length);
+            byteArrayFeeder.endOfInput();
+
+            _asyncParserFailsTooLongNumber(p, JsonToken.VALUE_NUMBER_FLOAT);
+        }
+    }
+    
+    private void _asyncParserFailsTooLongNumber(JsonParser p, JsonToken tokenMatch) throws Exception {
+        boolean foundNumber = false;
+        try {
+            while (p.nextToken() != null) {
+                if (p.currentToken() == tokenMatch) {
+                    foundNumber = true;
+                    String numberText = p.getText();
+                    assertEquals(TEST_NUMBER_LENGTH, numberText.length(),
+                            "Async parser silently accepted all " + TEST_NUMBER_LENGTH + " digits");
+                }
+            }
+            fail("Async parser must reject a " + TEST_NUMBER_LENGTH + "-digit number (number found? "+foundNumber+")");
+        } catch (StreamConstraintsException e) {
+            verifyException(e, "Number value length (");
+            verifyException(e, " exceeds the maximum allowed");
+        }
+    }
+
+    private byte[] buildPayloadWithLongInteger(int numDigits) {
+        StringBuilder sb = new StringBuilder(numDigits + 10);
+        sb.append("{\"v\":");
+        for (int i = 0; i < numDigits; i++) {
+            sb.append((char) ('1' + (i % 9)));
+        }
+        sb.append('}');
+        return utf8Bytes(sb.toString());
+    }
+
+    private byte[] buildPayloadWithLongDecimal(int numDigits) {
+        StringBuilder sb = new StringBuilder(numDigits + 10);
+        sb.append("{\"v\":0.");
+        for (int i = 0; i < numDigits; i++) {
+            sb.append((char) ('1' + (i % 9)));
+        }
+        sb.append('}');
+        return utf8Bytes(sb.toString());
+    }
+
+    private byte[] buildPayloadWithLongExponent(int numDigits) {
+        StringBuilder sb = new StringBuilder(numDigits + 10);
+        sb.append("{\"v\":1.1E");
+        for (int i = 0; i < numDigits; i++) {
+            sb.append((char) ('1' + (i % 9)));
+        }
+        return utf8Bytes(sb.toString());
+    }
+}
